Move utilities to utils package
