+++ app/src/main/java/io/github/vibrouter/MainService.java
@@ -35,6 +35,8 @@ import com.google.gson.GsonBuilder;
 import java.util.Collections;
 import java.util.List;
 
+import io.github.vibrouter.utils.GpsUtil;
+
 public class MainService extends Service {
     public interface CurrentLocationListener {
         void onLocationChanged(LatLng location);
