+++ app/src/main/java/io/github/vibrouter/utils/GpsUtil.java
@@ -1,4 +1,4 @@
-package io.github.vibrouter;
+package io.github.vibrouter.utils;
 
 import com.google.android.gms.maps.model.LatLng;
